moved plugins to public folder
