+++ packages/cli-kit/src/public/node/plugins/tunnel.ts
@@ -1,6 +1,6 @@
 import {err, Result} from '../result.js'
-import {FanoutHookFunction, PluginReturnsForHook} from '../../../plugins.js'
 import {ExtendableError} from '../../../error.js'
+import {FanoutHookFunction, PluginReturnsForHook} from '../plugins.js'
 
 export type TunnelErrorType = 'invalid-provider' | 'tunnel-already-running' | 'wrong-credentials' | 'unknown'
 export class TunnelError extends ExtendableError {
